requirements.txt
streamlit
pandas
numpy
plotly
fpdf2
camelot-py[cv]
tabula-py
python-dotenv